pyproject.toml
[project]
name = "payment-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "stripe[async]>=14.1.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
